requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
cffi==1.12.2
Click==7.0
coverage==4.5.3
cryptography==2.6.1
Flask==1.0.2
Flask-PyMongo==2.2.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==6.0.0
pathlib2==2.3.3
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.7.2
PyMySQL==0.9.3
pytest==4.3.1
six==1.12.0
SQLAlchemy==1.3.1
typed-ast==1.3.1
Werkzeug==0.15.1
wrapt==1.11.1
